# packages/api/internal/api/docs/swagger.yaml
basePath: /v1
definitions:
  domain.ExtractField:
    properties:
      attr:
        description: '"text" (default), "html", or attribute name'
        type: string
      multiple:
        type: boolean
      selector:
        type: string
    type: object
  domain.ImageData:
    properties:
      alt:
        type: string
      blob:
        type: string
      format:
        type: string
      height:
        type: integer
      size_bytes:
        type: integer
      source:
        type: string
      title:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
  domain.LinkData:
    properties:
      isInternal:
        type: boolean
      text:
        type: string
      url:
        type: string
    type: object
  domain.ScrapeResult:
    properties:
      extracted:
        additionalProperties: {}
        description: Extracted holds deterministic schema-extraction results.
        type: object
      html:
        type: string
      images:
        items:
          $ref: "#/definitions/domain.ImageData"
        type: array
      links:
        description:
          Links holds extracted hyperlinks (after readability, deduped,
          absolute).
        items:
          $ref: "#/definitions/domain.LinkData"
        type: array
      markdown:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      screenshot:
        allOf:
          - $ref: "#/definitions/domain.ScreenshotData"
        description: Image fields (omitted when not requested)
      summary:
        description: Summary holds the extractive summary when requested.
        type: string
      url:
        type: string
    type: object
  domain.ScreenshotData:
    properties:
      blob:
        type: string
      captured_at:
        type: string
      format:
        type: string
      full_page:
        type: boolean
      height:
        type: integer
      size_bytes:
        type: integer
      url:
        type: string
      width:
        type: integer
    type: object
  handlers.ActionReq:
    properties:
      ms:
        type: integer
      selector:
        type: string
      type:
        type: string
    type: object
  handlers.BatchRequest:
    properties:
      urls:
        items:
          type: string
        maxItems: 20
        minItems: 1
        type: array
    required:
      - urls
    type: object
  handlers.BatchResponse:
    properties:
      batch_id:
        type: string
      tasks:
        items:
          $ref: "#/definitions/handlers.BatchTask"
        type: array
    type: object
  handlers.BatchStatusResponse:
    properties:
      batch_id:
        type: string
      completed:
        type: integer
      failed:
        type: integer
      tasks:
        items:
          $ref: "#/definitions/handlers.BatchTask"
        type: array
      total:
        type: integer
    type: object
  handlers.BatchTask:
    properties:
      id:
        type: string
      url:
        type: string
    type: object
  handlers.CrawlRequest:
    properties:
      exclude_paths:
        items:
          type: string
        type: array
      image_format:
        type: string
      images:
        type: boolean
      include_paths:
        items:
          type: string
        type: array
      limit:
        type: integer
      maxDepth:
        type: integer
      render:
        type: boolean
      screenshot:
        type: boolean
      url:
        type: string
      webhook_secret:
        type: string
      webhook_url:
        type: string
    required:
      - url
    type: object
  handlers.CrawlResponse:
    properties:
      id:
        type: string
      image_format:
        type: string
      images:
        type: boolean
      limit:
        type: integer
      maxDepth:
        type: integer
      render:
        type: boolean
      screenshot:
        type: boolean
      url:
        type: string
    type: object
  handlers.ImageProcessReq:
    properties:
      format:
        description: '"jpeg" (default) or "png"'
        type: string
      max_width:
        type: integer
      quality:
        type: integer
    type: object
  handlers.MapRequest:
    properties:
      limit:
        description: default 100, max 5000
        type: integer
      search:
        description: optional substring filter
        type: string
      url:
        type: string
    required:
      - url
    type: object
  handlers.MapResponse:
    properties:
      count:
        type: integer
      links:
        items:
          $ref: "#/definitions/sitemap.DiscoveredURL"
        type: array
      url:
        type: string
    type: object
  handlers.MonitorRequest:
    properties:
      interval_seconds:
        type: integer
      url:
        type: string
      webhook_secret:
        type: string
      webhook_url:
        type: string
    required:
      - url
    type: object
  handlers.MonitorResponse:
    properties:
      id:
        type: string
      interval_seconds:
        type: integer
      next_check:
        type: string
      url:
        type: string
    type: object
  handlers.MultiScrapeItem:
    properties:
      error:
        type: string
      extracted:
        additionalProperties: {}
        type: object
      html:
        type: string
      images:
        items:
          $ref: "#/definitions/domain.ImageData"
        type: array
      links:
        items:
          $ref: "#/definitions/domain.LinkData"
        type: array
      markdown:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      screenshot:
        $ref: "#/definitions/domain.ScreenshotData"
      summary:
        type: string
      title:
        type: string
      url:
        type: string
      word_count:
        type: integer
    type: object
  handlers.MultiScrapeResponse:
    properties:
      results:
        items:
          $ref: "#/definitions/handlers.MultiScrapeItem"
        type: array
    type: object
  handlers.ScrapeRequest:
    properties:
      actions:
        items:
          $ref: "#/definitions/handlers.ActionReq"
        type: array
      block_ads:
        type: boolean
      extract_schema:
        additionalProperties:
          $ref: "#/definitions/domain.ExtractField"
        type: object
      image_format:
        description: '"url" or "blob"'
        type: string
      image_process:
        $ref: "#/definitions/handlers.ImageProcessReq"
      images:
        type: boolean
      include_links:
        type: boolean
      max_image_size_kb:
        type: integer
      max_images:
        type: integer
      mode:
        description: '"smart", "static", "dynamic"'
        type: string
      redact_pii:
        type: boolean
      remove_base64_images:
        type: boolean
      render:
        description: "Deprecated: usage ignores Mode if true"
        type: boolean
      screenshot:
        type: boolean
      screenshot_opts:
        $ref: "#/definitions/handlers.ScreenshotOpts"
      summary:
        type: boolean
      summary_sentences:
        type: integer
      url:
        type: string
      urls:
        items:
          type: string
        type: array
    type: object
  handlers.ScreenshotOpts:
    properties:
      format:
        type: string
      full_page:
        type: boolean
      height:
        type: integer
      quality:
        type: integer
      wait_selector:
        type: string
      width:
        type: integer
    type: object
  handlers.SearchRequest:
    properties:
      category:
        type: string
      excludeDomains:
        items:
          type: string
        type: array
      includeDomains:
        items:
          type: string
        type: array
      limit:
        type: integer
      maxAge:
        type: integer
      mode:
        type: string
      offset:
        type: integer
      query:
        type: string
      requiredText:
        items:
          type: string
        type: array
      rerank:
        type: boolean
    required:
      - query
    type: object
  handlers.SearchResponse:
    properties:
      count:
        type: integer
      hasMore:
        type: boolean
      nextOffset:
        type: integer
      query:
        type: string
      results:
        items:
          $ref: "#/definitions/search.Result"
        type: array
    type: object
  search.Result:
    properties:
      description:
        type: string
      domain:
        type: string
      highlights:
        items:
          type: string
        type: array
      id:
        type: string
      relevance:
        type: number
      title:
        type: string
      url:
        type: string
    type: object
  sitemap.DiscoveredURL:
    properties:
      source:
        description: '"sitemap" or "link"'
        type: string
      url:
        type: string
    type: object
info:
  contact: {}
  description: Web scraping, crawling, and AI data extraction API.
  title: Tomoshibi API
  version: "1.0"
paths:
  /batch/{id}:
    get:
      description: Aggregates the Asynq state of every task in the batch.
      parameters:
        - description: The batch ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handlers.BatchStatusResponse"
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get the status of a batch scrape
      tags:
        - batch
  /batch/scrape:
    post:
      consumes:
        - application/json
      description:
        Enqueues one scrape task per URL and returns a batch ID for status
        polling.
      parameters:
        - description: JSON request body
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/handlers.BatchRequest"
      produces:
        - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: "#/definitions/handlers.BatchResponse"
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Scrape multiple URLs asynchronously
      tags:
        - batch
  /crawl:
    post:
      consumes:
        - application/json
      description:
        Submits a URL to be crawled asynchronously. The crawler performs
        BFS link-following up to maxDepth, scraping up to limit pages on the same
        domain. Requires Redis.
      parameters:
        - description: JSON request body
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/handlers.CrawlRequest"
      produces:
        - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: "#/definitions/handlers.CrawlResponse"
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Enqueue a URL for asynchronous crawling
      tags:
        - crawl
  /crawl/{id}:
    get:
      description:
        Retrieves the current status and result of a previously enqueued
        crawl task by its ID.
      parameters:
        - description: The crawl task ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: The task status and result payload
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties: true
            type: object
      summary: Get the status of an asynchronous crawl
      tags:
        - crawl
  /map:
    post:
      consumes:
        - application/json
      description:
        Discovers URLs for a site via robots.txt/sitemap.xml, falling back
        to link discovery. Returns up to limit URLs, optionally filtered by search.
      parameters:
        - description: JSON request body
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/handlers.MapRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handlers.MapResponse"
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Map a website's URLs
      tags:
        - map
  /monitor:
    post:
      consumes:
        - application/json
      description:
        Scrapes the URL on an interval, hashes the markdown, and fires
        a webhook when content changes. First check records the baseline.
      parameters:
        - description: JSON request body
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/handlers.MonitorRequest"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/handlers.MonitorResponse"
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a change-tracking monitor
      tags:
        - monitor
  /monitor/{id}:
    delete:
      description: Stops monitoring and removes the monitor record.
      parameters:
        - description: The monitor ID
          in: path
          name: id
          required: true
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a monitor
      tags:
        - monitor
    get:
      description:
        Returns the monitor configuration, last content hash, and next
        check time.
      parameters:
        - description: The monitor ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get monitor status
      tags:
        - monitor
  /scrape:
    get:
      consumes:
        - application/json
      description:
        'Scrapes a given URL and returns its markdown content, metadata,
        links, and optionally captures a screenshot or extracts images if enabled.
        Links are extracted from readability DOM, resolved to absolute URLs, deduped,
        and flagged isInternal via same-host check — matching Firecrawl `formats:
        ["links"]` (Firecrawl returns `links: ["https://..."]`; Cinder enriches to
        `links: [{url, text, isInternal}]`). Include_links defaults to true. Accepts
        either `url` (single) or `urls` (up to 10) for synchronous multi-URL scrape;
        when `urls` is used the response is `{results: [{url, markdown, metadata,
        ...}]}`. Mirrors Firecrawl `POST /v2/scrape` (single) and `POST /v2/batch/scrape`
        (multi) and the `web_fetch_exa` multi-URL shape — see https://docs.firecrawl.dev/features/scrape
        and https://docs.firecrawl.dev/api-reference/endpoint/scrape. Research via
        http://localhost:3002/v1/scrape against those docs showed `links` as string
        array; Cinder returns objects with text/isInternal.'
      parameters:
        - description: The URL to scrape (single mode)
          in: query
          name: url
          type: string
        - description: 'JSON array alternative: use POST body {\'
          in: query
          name: urls
          type: string
        - description: "Scraping mode: smart, static, dynamic"
          in: query
          name: mode
          type: string
        - description: "Deprecated: use mode=dynamic instead"
          in: query
          name: render
          type: boolean
        - description: Capture full-page screenshot (requires mode=dynamic or smart)
          in: query
          name: screenshot
          type: boolean
        - description: Extract images from the page
          in: query
          name: images
          type: boolean
        - description:
            "Image transport: 'url' (default, metadata only) or 'blob'
            (base64 data URIs)"
          in: query
          name: image_format
          type: string
        - description: "Maximum images to extract (default: 10)"
          in: query
          name: max_images
          type: integer
        - description: "Max individual image size in KB (default: 5120)"
          in: query
          name: max_image_size_kb
          type: integer
        - description: Include extracted links (default true)
          in: query
          name: include_links
          type: boolean
        - description:
            JSON request body (alternative to query params). Provide either
            `url` or `urls` (max 10, exclusive).
          in: body
          name: body
          schema:
            $ref: "#/definitions/handlers.ScrapeRequest"
      produces:
        - application/json
      responses:
        "200":
          description: multi URL (when urls is provided)
          schema:
            $ref: "#/definitions/handlers.MultiScrapeResponse"
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Scrape a webpage
      tags:
        - scrape
    post:
      consumes:
        - application/json
      description:
        'Scrapes a given URL and returns its markdown content, metadata,
        links, and optionally captures a screenshot or extracts images if enabled.
        Links are extracted from readability DOM, resolved to absolute URLs, deduped,
        and flagged isInternal via same-host check — matching Firecrawl `formats:
        ["links"]` (Firecrawl returns `links: ["https://..."]`; Cinder enriches to
        `links: [{url, text, isInternal}]`). Include_links defaults to true. Accepts
        either `url` (single) or `urls` (up to 10) for synchronous multi-URL scrape;
        when `urls` is used the response is `{results: [{url, markdown, metadata,
        ...}]}`. Mirrors Firecrawl `POST /v2/scrape` (single) and `POST /v2/batch/scrape`
        (multi) and the `web_fetch_exa` multi-URL shape — see https://docs.firecrawl.dev/features/scrape
        and https://docs.firecrawl.dev/api-reference/endpoint/scrape. Research via
        http://localhost:3002/v1/scrape against those docs showed `links` as string
        array; Cinder returns objects with text/isInternal.'
      parameters:
        - description: The URL to scrape (single mode)
          in: query
          name: url
          type: string
        - description: 'JSON array alternative: use POST body {\'
          in: query
          name: urls
          type: string
        - description: "Scraping mode: smart, static, dynamic"
          in: query
          name: mode
          type: string
        - description: "Deprecated: use mode=dynamic instead"
          in: query
          name: render
          type: boolean
        - description: Capture full-page screenshot (requires mode=dynamic or smart)
          in: query
          name: screenshot
          type: boolean
        - description: Extract images from the page
          in: query
          name: images
          type: boolean
        - description:
            "Image transport: 'url' (default, metadata only) or 'blob'
            (base64 data URIs)"
          in: query
          name: image_format
          type: string
        - description: "Maximum images to extract (default: 10)"
          in: query
          name: max_images
          type: integer
        - description: "Max individual image size in KB (default: 5120)"
          in: query
          name: max_image_size_kb
          type: integer
        - description: Include extracted links (default true)
          in: query
          name: include_links
          type: boolean
        - description:
            JSON request body (alternative to query params). Provide either
            `url` or `urls` (max 10, exclusive).
          in: body
          name: body
          schema:
            $ref: "#/definitions/handlers.ScrapeRequest"
      produces:
        - application/json
      responses:
        "200":
          description: multi URL (when urls is provided)
          schema:
            $ref: "#/definitions/handlers.MultiScrapeResponse"
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Scrape a webpage
      tags:
        - scrape
  /search:
    get:
      consumes:
        - application/json
      description:
        Searches the web using DuckDuckGo HTML (free, no key) with Brave
        Search as fallback when BRAVE_SEARCH_API_KEY is set.
      parameters:
        - description: The search query
          in: query
          name: query
          type: string
        - description: Alias for query
          in: query
          name: q
          type: string
        - description: Pagination offset
          in: query
          name: offset
          type: integer
        - description: Pagination limit (max 100)
          in: query
          name: limit
          type: integer
        - description: "Category filter: general, news, code"
          enum:
            - general
            - news
            - code
          in: query
          name: category
          type: string
        - description: Lightweight TF-IDF re-rank (pure Go, no ONNX)
          in: query
          name: rerank
          type: boolean
        - description: JSON request body
          in: body
          name: body
          schema:
            $ref: "#/definitions/handlers.SearchRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handlers.SearchResponse"
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Search the web
      tags:
        - search
    post:
      consumes:
        - application/json
      description:
        Searches the web using DuckDuckGo HTML (free, no key) with Brave
        Search as fallback when BRAVE_SEARCH_API_KEY is set.
      parameters:
        - description: The search query
          in: query
          name: query
          type: string
        - description: Alias for query
          in: query
          name: q
          type: string
        - description: Pagination offset
          in: query
          name: offset
          type: integer
        - description: Pagination limit (max 100)
          in: query
          name: limit
          type: integer
        - description: "Category filter: general, news, code"
          enum:
            - general
            - news
            - code
          in: query
          name: category
          type: string
        - description: Lightweight TF-IDF re-rank (pure Go, no ONNX)
          in: query
          name: rerank
          type: boolean
        - description: JSON request body
          in: body
          name: body
          schema:
            $ref: "#/definitions/handlers.SearchRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handlers.SearchResponse"
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Search the web
      tags:
        - search
swagger: "2.0"